[IMPROVE] adjust invoice entity to store data based on midtrans notification request and add IP whitelist filtering method to secure the midtrans notification endpoint

+++ src/main/java/io/github/muhammadredin/tokonyadiaapi/dto/response/ProductOrderResponse.java
@@ -10,6 +10,6 @@ import lombok.*;
 public class ProductOrderResponse {
     private String productId;
     private String productName;
-    private Integer productPrice;
+    private Long productPrice;
     private Integer quantity;
 }
